pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name        = "constitutional-os"
version     = "0.1.0"
description = "A formal runtime for epistemic-governance systems: profiles, invariants, membranes, delta calculus, Lyapunov stability, and A-safety proofs."
readme      = "README.md"
license     = { text = "Apache-2.0" }
requires-python = ">=3.10"

keywords = [
    "ai-safety", "governance", "formal-methods",
    "attractor", "lyapunov", "constitutional-ai",
    "reliability", "invariants", "membranes",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

# Core dependencies — intentionally minimal
dependencies = []

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
api  = ["fastapi>=0.110", "uvicorn[standard]>=0.29"]
viz  = ["matplotlib>=3.7"]
all  = ["pyyaml>=6.0", "fastapi>=0.110", "uvicorn[standard]>=0.29", "matplotlib>=3.7"]
dev  = ["pytest>=7.0", "pytest-cov>=4.0", "mypy>=1.0", "ruff>=0.1"]

[project.scripts]
constitutional-os = "constitutional_os.console.cli:main"

[project.urls]
Homepage    = "https://github.com/jazzybird07/constitutional-os"
Repository  = "https://github.com/jazzybird07/constitutional-os"
"Bug Tracker" = "https://github.com/jazzybird07/constitutional-os/issues"
Zenodo      = "https://zenodo.org/records/19045723"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
constitutional_os = ["py.typed"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version  = "3.10"
strict          = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]
ignore = ["E501"]
